chatgpt/requirements.txt
anyio==3.6.2
blobfile==2.0.1
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
filelock==3.9.0
firebase-admin==6.1.0
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-pubsub==2.15.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpc-google-iam-v1==0.12.6
grpcio==1.51.3
grpcio-status==1.51.3
h11==0.14.0
httpcore==0.16.3
httplib2==0.21.0
httpx==0.23.3
idna==3.4
lxml==4.9.2
msgpack==1.0.5
OpenAIAuth==0.3.2
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.17
PyJWT==2.6.0
pyparsing==3.0.9
PySocks==1.7.1
regex==2022.10.31
requests==2.28.2
revChatGPT==3.1.5
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
tiktoken==0.3.0
uritemplate==4.1.1
urllib3==1.26.14
wcwidth==0.2.6
